EA-162 Move the rest of RAML 0.8 to RAML 1.0

Approved-by: Andrew Knight <avenak.consulting@gmail.com>

+++ mobile-web/src/test/java/net/syscon/elite/executableSpecification/PrisonerSearchStepDefinitions.java
@@ -57,7 +57,7 @@ public class PrisonerSearchStepDefinitions extends AbstractStepDefinitions {
     }
 
     @When("^a search is made for prisoners with DOB on or after (\\d+-\\d+-\\d+) for range ([0-9]*) -> ([0-9]*)$")
-    public void aSearchIsMadeForPrisonersWithDOBOnOrAfterForRange(String dobFrom, int offset, int limit) throws Throwable {
+    public void aSearchIsMadeForPrisonersWithDOBOnOrAfterForRange(String dobFrom, long offset, long limit) throws Throwable {
         prisonerSearch.search(ImmutableMap.of("dobFrom", dobFrom), offset, limit, HttpStatus.OK);
     }
     @When("^a search is made for prisoners with DOB between \"([^\"]*)\" and \"([^\"]*)\" for range ([0-9]*) -> ([0-9]*)$")
